fix(web): hide broken page actions on docs index pages
